feat: created separate components for each cardType
