+++ src/ProfileCards/App.tsx
@@ -13,13 +13,13 @@ import { useNavigate } from "react-router-dom";
 import {
 	convertEpochToISO,
 	getDateAndMonth,
-	getDaysAgo,
 	parseUrl,
 	replaceSlugInURL,
 	updateCardTypeToRenderInAllCards,
 } from "../common/utils";
-import { YTCardBody } from "./YTCardBody";
-import { ImageCard } from "./ImageCard";
+import { DataCard } from "./CardTypeComponents/DataCard";
+import { YTCard } from "./CardTypeComponents/YTCard";
+import { ImageCard } from "./CardTypeComponents/ImageCard";
 
 const GET_CARD_DETAIL = "cards/pending/{slug}";
 const CREATE_PUBLISHED_CARDS = "cards/published/{slug}";
@@ -124,143 +124,24 @@ export default function App({ user, setUser, slug }: UserDataProps) {
 						{/* Actual Card with all Data */}
 						<div className="flex flex-col md:flex-row justify-between items-center mb-[16px] mt-[16px]">
 							<div className="flex flex-col md:flex-row justify-center items-stretch gap-4 mx-auto">
-								{/* CardType == DATA CARD */}
-								{activeCard.cardType == "DataCard" && (
-									<div
-										className={`rounded-lg shadow-lg p-3 px-5 flex flex-col justify-between min-h-[desiredMinHeight] gap-2 ${
-											activeCard.cardTypeToRender === CardTypeToRender.YT
-												? "bg-yt-card max-h-[230px]"
-												: "bg-white max-h-[250px]"
-										}`}
-										style={
-											activeCard.cardTypeToRender === CardTypeToRender.PURPLE
-												? {
-														backgroundImage: `url('../assets/images/purpleCardBg.png')`,
-												  }
-												: {}
-										}
-									>
-										{/* Body for YT card */}
-										{activeCard.cardTypeToRender == CardTypeToRender.YT && (
-											<YTCardBody card={activeCard} />
-										)}
-
-										{/* Header with favicons and date. */}
-										<header className="relative flex items-center">
-											{/* Map over all urls and show the favicon */}
-											{Array.from(
-												new Set(
-													activeCard.urls.map(
-														(url) =>
-															`https://www.google.com/s2/favicons?domain=${parseUrl(
-																url.url
-															)}&sz=40`
-													)
-												)
-											).map((iconUrl, index) => (
-												<div
-													key={iconUrl}
-													className="w-6 h-6 flex-none rounded-full border-spacing-4"
-												>
-													<img
-														className={`absolute w-6 h-6 flex-none rounded-full border-2 ${
-															activeCard.cardTypeToRender ===
-															CardTypeToRender.PURPLE
-																? "border-purple-card fill-purple-card"
-																: "border-white fill-white"
-														}`}
-														style={{
-															left: `${index * 1.3}rem`,
-														}}
-														src={iconUrl}
-													/>
-												</div>
-											))}
-											<div className="flex flex-row ml-auto items-center">
-												<div
-													className={`flex font-inter text-sm ${
-														activeCard.cardTypeToRender ===
-														CardTypeToRender.PURPLE
-															? "text-white"
-															: "text-gray-400"
-													} font-normal`}
-												>
-													{getDaysAgo(activeCard.date)}
-												</div>
-											</div>
-										</header>
-
-										{/* Card Content */}
-										<div className="flex flex-col justify-center mt-1">
-											<blockquote
-												className={`text-base font-normal ${
-													activeCard.cardTypeToRender === CardTypeToRender.YT ||
-													activeCard.cardTypeToRender ===
-														CardTypeToRender.PURPLE
-														? "text-white"
-														: "text-gray-600"
-												}`}
-											>
-												{(activeCard.content.length > 60 &&
-													activeCard.cardTypeToRender ===
-														CardTypeToRender.YT) ||
-												(activeCard.content.length > 120 &&
-													activeCard.cardTypeToRender !== CardTypeToRender.YT)
-													? `${activeCard.content.slice(
-															0,
-															activeCard.cardTypeToRender ===
-																CardTypeToRender.YT
-																? 60
-																: 120
-													  )}...`
-													: activeCard.content}
-											</blockquote>
-										</div>
-
-										{/* URL pills in bottom */}
-										{(activeCard.cardTypeToRender == CardTypeToRender.DATA ||
-											activeCard.cardTypeToRender ==
-												CardTypeToRender.PURPLE) && (
-											<div className="flex flex-row w-full flex-wrap gap-2 self-stretch items-center justify-start pt-5">
-												<>
-													{activeCard.urls.map((urls) => (
-														<button
-															className={`flex items-center gap-2 rounded-3xl border px-2 py-1 ${
-																activeCard.cardTypeToRender ==
-																CardTypeToRender.PURPLE
-																	? "border-none bg-white/20"
-																	: "border-gray-200 bg-gray-50"
-															}`}
-															onClick={() => handleOnClick(urls.url)}
-														>
-															<img
-																className="w-4 h-4 flex-none rounded-full"
-																src={`https://www.google.com/s2/favicons?domain=${urls.url}&sz=16`}
-															/>
-
-															<h3
-																className={`inline-block text-xs font-medium ${
-																	activeCard.cardTypeToRender ==
-																	CardTypeToRender.PURPLE
-																		? "text-white"
-																		: "text-gray-700"
-																} overflow-hidden overflow-ellipsis line-clamp-1`}
-															>
-																{activeCard.urls.length > 2 &&
-																urls.title.length > 10
-																	? urls.title.trim().slice(0, 10) + "..."
-																	: urls.title.trim().slice(0, 25) + "..."}
-															</h3>
-														</button>
-													))}
-												</>
-											</div>
-										)}
-									</div>
+								{/* CardType == DataCard || PurpleCard */}
+								{(activeCard.cardTypeToRender === CardTypeToRender.DATA ||
+									activeCard.cardTypeToRender === CardTypeToRender.PURPLE) && (
+									<DataCard
+										activeCard={activeCard}
+										handleOnClick={handleOnClick}
+										key={activeCard.id}
+									/>
 								)}
 
-								{activeCard.cardType == "ImageCard" && (
+								{/* CardType == ImageCard */}
+								{activeCard.cardTypeToRender === CardTypeToRender.IMAGE && (
 									<ImageCard card={activeCard} />
+								)}
+
+								{/* CardType == YTCard */}
+								{activeCard.cardTypeToRender === CardTypeToRender.YT && (
+									<YTCard activeCard={activeCard} />
 								)}
 
 								{/* CardType == DOMAIN VISITED CARD */}
